# v6: preprocess config for evaluation