--- a/v6/config_parser.go
+++ b/v6/config_parser.go
@@ -1,11 +1,8 @@
 package configcat
 
 import (
-	"crypto/sha1"
-	"encoding/binary"
 	"encoding/json"
 	"fmt"
-	"strings"
 	"time"
 )
 
@@ -21,6 +18,9 @@ type config struct {
 	jsonBody  string
 	etag      string
 	root      *rootNode
+	evaluate  func(logger Logger, key string, user *User) (interface{}, string, error)
+	allKeys   []string
+	keyValues map[string]keyValue
 	fetchTime time.Time
 }
 
@@ -32,6 +32,9 @@ func parseConfig(jsonBody []byte, etag string, fetchTime time.Time) (*config, er
 	return &config{
 		jsonBody:  string(jsonBody),
 		root:      &root,
+		evaluate:  evaluator(&root),
+		keyValues: keyValuesForRootNode(&root),
+		allKeys:   keysForRootNode(&root),
 		etag:      etag,
 		fetchTime: fetchTime,
 	}, nil
@@ -51,128 +54,19 @@ func (c *config) body() string {
 }
 
 func (conf *config) getKeyAndValueForVariation(variationId string) (string, interface{}) {
-	for key, entry := range conf.root.Entries {
-		if entry.VariationID == variationId {
-			return key, entry.Value
-		}
-		for _, rule := range entry.RolloutRules {
-			if rule.VariationID == variationId {
-				return key, rule.Value
-			}
-		}
-		for _, rule := range entry.PercentageRules {
-			if rule.VariationID == variationId {
-				return key, rule.Value
-			}
-		}
-	}
-	return "", nil
+	kv := conf.keyValues[variationId]
+	return kv.key, kv.value
 }
 
 func (conf *config) getAllKeys() []string {
-	keys := make([]string, 0, len(conf.root.Entries))
-	for k := range conf.root.Entries {
-		keys = append(keys, k)
-	}
-	return keys
+	return conf.allKeys
 }
 
 func (conf *config) getValueAndVariationId(logger Logger, key string, user *User) (interface{}, string, error) {
-	if len(key) == 0 {
-		return nil, "", fmt.Errorf("key cannot be empty")
-	}
 	if conf == nil {
 		return nil, "", fmt.Errorf("no configuration available")
 	}
-	entryNode := conf.root.Entries[key]
-	if entryNode == nil {
-		return nil, "", &parseError{
-			"Value not found for key " + key +
-				". Here are the available keys: " + strings.Join(conf.getAllKeys(), ", "),
-		}
-	}
-
-	parsed, variationId := conf.evaluate(logger, entryNode, key, user)
-	if parsed == nil {
-		return nil, "", &parseError{"Null evaluated for key " + key + "."}
-	}
-
-	return parsed, variationId, nil
-}
-
-func (conf *config) evaluate(logger Logger, node *entry, key string, user *User) (interface{}, string) {
-	logger.Infof("Evaluating GetValue(%s).", key)
-
-	if user == nil {
-		if len(node.RolloutRules) > 0 || len(node.PercentageRules) > 0 {
-			logger.Warnln("Evaluating GetValue(" + key + "). UserObject missing! You should pass a " +
-				"UserObject to GetValueForUser() in order to make targeting work properly. " +
-				"Read more: https://configcat.com/docs/advanced/user-object.")
-		}
-
-		result := node.Value
-		logger.Infof("Returning %v.", result)
-		return result, node.VariationID
-	}
-
-	logger.Infof("User object: %v", user)
-
-	for _, rule := range node.RolloutRules {
-		if rule == nil {
-			continue
-		}
-		matched, err := matchRolloutRule(rule, user)
-		if err != nil {
-			logger.Infof("Evaluating rule: [%s:%s] [%s] [%s] => SKIP rule. Validation error: %v",
-				rule.ComparisonAttribute,
-				user,
-				rule.Comparator,
-				rule.Value,
-				err,
-			)
-			continue
-		}
-		if matched {
-			logger.Infof("Evaluating rule: [%s:%s] [%s] [%s] => match, returning: %v",
-				rule.ComparisonAttribute,
-				user,
-				rule.Comparator,
-				rule.ComparisonValue,
-				rule.Value,
-			)
-			return rule.Value, rule.VariationID
-		}
-		logger.Infof("Evaluating rule: [%s:%s] [%s] [%s] => no match",
-			rule.ComparisonAttribute,
-			user,
-			rule.Comparator,
-			rule.Value,
-		)
-	}
-
-	if len(node.PercentageRules) > 0 {
-		sum := sha1.Sum([]byte(key + user.identifier))
-		// Treat the first 4 bytes as a number, then knock
-		// of the last 4 bits. This is equivalent to turning the
-		// entire sum into hex, then decoding the first 7 digits.
-		num := int64(binary.BigEndian.Uint32(sum[:4]))
-		num >>= 4
-
-		scaled := num % 100
-		bucket := int64(0)
-		for _, rule := range node.PercentageRules {
-			bucket += rule.Percentage
-			if scaled < bucket {
-				result := rule.Value
-				logger.Infof("Evaluating %% options. Returning %s", result)
-				return result, rule.VariationID
-			}
-		}
-	}
-
-	result := node.Value
-	logger.Infof("Returning %v.", result)
-	return result, node.VariationID
+	return conf.evaluate(logger, key, user)
 }
 
 type rootNode struct {