Instead of evaluating the config node on each request,
we preprocess it into a more efficient form.

```
benchmark               old ns/op     new ns/op     delta
BenchmarkGetValue-8     467           82.5          -82.34%

benchmark               old allocs     new allocs     delta
BenchmarkGetValue-8     6              0              -100.00%

benchmark               old bytes     new bytes     delta
BenchmarkGetValue-8     304           0             -100.00%
```